Start kanshi from sway config and SIGHUP on config reload
I tried to run kanshi as a guix service for a while, but I never quite
got that working as well as I'd like it to. Switching to starting
kanshi from the sway config (which ensures it always starts after sway
does) and reloading it on a sway config reload.